Tweak server settings tests
- move testMigrate to LegacyServerSettingsTest
- fix conflicts between tests
