+++ dicoogle/src/test/java/pt/ua/dicoogle/core/settings/LegacyServerSettingsTest.java
@@ -18,18 +18,22 @@
  */
 package pt.ua.dicoogle.core.settings;
 
-import org.junit.Before;
 import org.junit.ComparisonFailure;
 import org.junit.Test;
 import pt.ua.dicoogle.sdk.datastructs.MoveDestination;
+import pt.ua.dicoogle.sdk.datastructs.SOPClass;
 import pt.ua.dicoogle.sdk.settings.server.ServerSettings;
+import pt.ua.dicoogle.server.SOPList;
 
 import java.io.IOException;
 import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.junit.Assert.*;
 
@@ -38,17 +42,12 @@ import static org.junit.Assert.*;
  */
 public class LegacyServerSettingsTest {
 
-    private URL testConfig;
-
-    @Before
-    public void init() {
-        this.testConfig = this.getClass().getResource("test-config-legacy.xml");
-    }
+    private final URL legacyConfig = this.getClass().getResource("test-config-legacy.xml");
 
     @Test
     public void test() throws IOException {
         // read the settings from our test config file
-        ServerSettings settings = ServerSettingsManager.loadLegacySettingsAt(this.testConfig);
+        ServerSettings settings = ServerSettingsManager.loadLegacySettingsAt(this.legacyConfig);
 
         assertTrue(settings instanceof LegacyServerSettings);
 
@@ -93,7 +92,7 @@ public class LegacyServerSettingsTest {
         assertSameContent(destinations, settings.getDicomServicesSettings().getMoveDestinations());
     }
 
-    private static void assertSameContent(Collection o1, Collection o2) {
+    private static void assertSameContent(Collection<?> o1, Collection<?> o2) {
         for (Object o : o1) {
             if (!o2.contains(o)) {
                 throw new ComparisonFailure("Collections do not have the same content", String.valueOf(o1),
@@ -108,4 +107,74 @@ public class LegacyServerSettingsTest {
         }
     }
 
+    @Test
+    public void testMigrate() throws IOException {
+        Path newconf = Files.createTempFile("conf", ".xml");
+
+        // load legacy server settings
+        ServerSettings settings = ServerSettingsManager.loadLegacySettingsAt(this.legacyConfig);
+        assertTrue(settings instanceof LegacyServerSettings);
+
+        // save as new
+        ServerSettingsManager.saveSettingsTo(settings, newconf);
+
+        // check new file
+        settings = ServerSettingsManager.loadSettingsAt(newconf);
+        assertTrue(settings instanceof ServerSettingsImpl);
+        ServerSettings.Archive a = settings.getArchiveSettings();
+
+        // assertions follow
+        assertEquals("/opt/dicoogle/repository", a.getMainDirectory());
+        assertEquals("/tmp", a.getWatchDirectory());
+        assertEquals(97, a.getIndexerEffort());
+        assertEquals("dicoogle-old", a.getNodeName());
+
+        assertEquals("TEST-STORAGE", settings.getDicomServicesSettings().getAETitle());
+
+        // QR settings
+        assertEquals(106, settings.getDicomServicesSettings().getQueryRetrieveSettings().getPort());
+        assertSameContent(Collections.singleton("any"),
+                settings.getDicomServicesSettings().getAllowedLocalInterfaces());
+        assertSameContent(Collections.singleton("any"), settings.getDicomServicesSettings().getAllowedHostnames());
+        assertEquals(3, settings.getDicomServicesSettings().getQueryRetrieveSettings().getRspDelay());
+        assertEquals(50, settings.getDicomServicesSettings().getQueryRetrieveSettings().getDIMSERspTimeout());
+        assertEquals(50, settings.getDicomServicesSettings().getQueryRetrieveSettings().getIdleTimeout());
+        assertEquals(50, settings.getDicomServicesSettings().getQueryRetrieveSettings().getAcceptTimeout());
+        assertEquals(50, settings.getDicomServicesSettings().getQueryRetrieveSettings().getConnectionTimeout());
+        // assertEquals("1.2.840.10008.5.1.4.1.2.1.1", settings.getSOPClasses());
+        assertEquals(22, settings.getDicomServicesSettings().getQueryRetrieveSettings().getMaxClientAssoc());
+        assertEquals(16360, settings.getDicomServicesSettings().getQueryRetrieveSettings().getMaxPDULengthSend());
+        assertEquals(16360, settings.getDicomServicesSettings().getQueryRetrieveSettings().getMaxPDULengthReceive());
+
+        // DICOM Storage settings
+        assertFalse(settings.getDicomServicesSettings().getStorageSettings().isAutostart());
+        assertEquals(6666, settings.getDicomServicesSettings().getStorageSettings().getPort());
+
+        // Web server settings
+        ServerSettings.WebServer web = settings.getWebServerSettings();
+        assertTrue(web.isAutostart());
+        assertEquals(8484, web.getPort());
+        assertEquals("test.dicoogle.com", web.getAllowedOrigins());
+
+
+        // SOP Classes
+
+        Collection<SOPClass> sopClasses = Arrays.asList(
+                new SOPClass("1.2.840.10008.5.1.4.1.1.88.40",
+                        Arrays.asList("1.2.840.10008.1.2", "1.2.840.10008.1.2.1", "1.2.840.10008.1.2.4.80",
+                                "1.2.840.10008.1.2.4.50")),
+                new SOPClass("1.2.840.10008.5.1.4.1.1.77.1.1",
+                        Arrays.asList("1.2.840.10008.1.2", "1.2.840.10008.1.2.1", "1.2.840.10008.1.2.4.80",
+                                "1.2.840.10008.1.2.4.50")),
+                new SOPClass("1.2.840.10008.5.1.1.30", Arrays.asList("1.2.840.10008.1.2", "1.2.840.10008.1.2.1",
+                        "1.2.840.10008.1.2.4.80", "1.2.840.10008.1.2.4.50")));
+
+        Collection<SOPClass> sopClassesFromSettings = settings.getDicomServicesSettings().getSOPClasses().stream()
+                .filter(c -> !c.getTransferSyntaxes().isEmpty()).collect(Collectors.toList());
+
+        assertSameContent(sopClasses, sopClassesFromSettings);
+
+        // clean up
+        Files.delete(newconf);
+    }
 }
